compile fix for Solaris SunPro CC
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@3480 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/qore_date_private.cpp
@@ -522,7 +522,7 @@ void qore_date_private::setRelativeDate(const char *str) {
    if (us) {
       l = strlen(p + 1);
       assert(l < 7);
-      us *= (int)pow(10, 6 - l);
+      us *= (int)pow((double)10, 6 - l);
    }         
 
    d.rel.setLiteral(date, us);
